# specs/api.yml
openapi: 3.0.0
info:
  title: OnePage 電商平台 API
  version: 1.0.0
  description: |
    OnePage 一頁式電商平台 API 契約。
    後端：Spring Boot 3.2 + JPA + PostgreSQL。
    金流：綠界全方位金流 API v5。
    發票：綠界電子發票 API。
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: 登入後台
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登入成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 帳號或密碼錯誤
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: 帳號已鎖定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: 使用 Refresh Token 換發 Access Token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: 換發成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Refresh Token 無效或已過期

  # ═══ User Management ═══
  /users:
    get:
      tags: [Auth]
      summary: 查詢使用者清單（管理員限定）
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 使用者清單
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        '403':
          description: 無權限

    post:
      tags: [Auth]
      summary: 建立使用者（管理員限定）（REQ-033）
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 使用者建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: 驗證失敗（Email 格式錯誤、密碼不足8碼等）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 無權限（僅管理員）
        '409':
          description: Email 已被使用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}/role:
    put:
      tags: [Auth]
      summary: 變更使用者角色（管理員限定）
      operationId: updateUserRole
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [ADMIN, GENERAL_USER]
      responses:
        '200':
          description: 角色變更成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: 不可將最後一位管理員降級
        '403':
          description: 無權限

  # ═══ Product Category ═══
...
  /invoices:
    get:
      tags: [Invoice]
      summary: 查詢發票清單
      operationId: listInvoices
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: query
          schema:
            type: integer
            format: int64
        - name: invoiceNumber
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [SYNCING, ISSUED, VOIDED, ALLOWANCED]
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 發票清單
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedInvoices'
  /invoices/{invoiceId}/void:
    post:
      tags: [Invoice]
      summary: 作廢發票
      operationId: voidInvoice
      security:
        - bearerAuth: []
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
                  description: 作廢原因
      responses:
        '200':
          description: 作廢成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDTO'
        '400':
          description: 僅已開立且同月份的發票可作廢
  /invoices/{invoiceId}/allowance:
    post:
      tags: [Invoice]
      summary: 折讓發票
      operationId: allowanceInvoice
      security:
        - bearerAuth: []
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  format: double
                  minimum: 0.01
                  description: 折讓金額（不超過原始發票金額）
      responses:
        '200':
          description: 折讓成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDTO'
        '400':
          description: 僅已開立發票可折讓或金額超過原始金額
  /invoices/{invoiceId}/sync:
    post:
      tags: [Invoice]
      summary: 手動重新同步發票（綠界）
      operationId: syncInvoice
      security:
        - bearerAuth: []
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 同步成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceDTO'
        '400':
          description: 僅同步中狀態的發票可手動同步

  # ═══ Inventory ═══
  /inventory:
    get:
      tags: [Inventory]
      summary: 查詢庫存清單
      operationId: listInventory
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 庫存清單
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryDTO'
  /inventory/{productId}:
    put:
      tags: [Inventory]
      summary: 編輯庫存數量
      operationId: updateInventory
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stockQuantity]
              properties:
                stockQuantity:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: 庫存更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryDTO'
        '400':
          description: 驗證失敗
  /inventory/{productId}/threshold:
    put:
      tags: [Inventory]
      summary: 設定低庫存門檻
      operationId: updateLowStockThreshold
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [threshold]
              properties:
                threshold:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: 門檻更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryDTO'
        '400':
          description: 驗證失敗

  # ═══ Email Template ═══
  /email-templates:
    get:
      tags: [EmailTemplate]
      summary: 查詢 Email 通知模板清單
      description: 返回當前登入使用者自己的四套模板（REQ-034 RBAC 資料隔離）。可用變數清單含中文說明（REQ-032）。
      operationId: listEmailTemplates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 模板清單（含可用變數說明）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmailTemplateDTO'
  /email-templates/{templateId}:
    put:
      tags: [EmailTemplate]
      summary: 編輯 Email 通知模板
      operationId: updateEmailTemplate
      security:
        - bearerAuth: []
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmailTemplateRequest'
      responses:
        '200':
          description: 模板更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailTemplateDTO'
        '400':
          description: 驗證失敗
        '403':
          description: 無權限（僅管理員）
  /email-templates/{templateId}/preview:
    get:
      tags: [EmailTemplate]
      summary: 預覽模板（使用範例資料）
      operationId: previewEmailTemplate
      security:
        - bearerAuth: []
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 預覽 HTML
          content:
            text/html:
              schema:
                type: string

  # ═══ System ═══
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        role:
          type: string
          enum: [ADMIN, GENERAL_USER]
        userName:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, GENERAL_USER]
    CreateUserRequest:
      type: object
      required: [email, name, password, role]
      description: 建立新使用者請求（REQ-033）
      properties:
        email:
          type: string
          format: email
          description: 登入 Email（唯一）
        name:
          type: string
          minLength: 1
          description: 使用者姓名
        password:
          type: string
          minLength: 8
          description: 密碼（8碼以上）
        role:
          type: string
          enum: [ADMIN, GENERAL_USER]
          description: 角色
        sendWelcomeEmail:
          type: boolean
          default: false
          description: 是否寄送歡迎 Email 給新使用者（REQ-033）

    # ── Product Category ──
...
    InvoiceDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        orderNumber:
          type: string
        invoiceNumber:
          type: string
          nullable: true
        randomCode:
          type: string
          nullable: true
        invoiceDate:
          type: string
          format: date
          nullable: true
        amount:
          type: number
          format: double
        invoiceType:
          type: string
          enum: [TWO_COPIES, THREE_COPIES]
        carrierType:
          type: string
          nullable: true
        carrierNumber:
          type: string
          nullable: true
        buyerTaxId:
          type: string
          nullable: true
        status:
          type: string
          enum: [SYNCING, ISSUED, VOIDED, ALLOWANCED]
        voidReason:
          type: string
          nullable: true
        allowanceAmount:
          type: number
          format: double
          nullable: true
        allowanceNumber:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PagedInvoices:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer

    # ── Inventory ──
    InventoryDTO:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        productName:
          type: string
        stockQuantity:
          type: integer
        lowStockThreshold:
          type: integer
        isLowStock:
          type: boolean

    # ── Email Template ──
    EmailTemplateDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        templateType:
          type: string
          enum: [ORDER_CONFIRMED, PAYMENT_SUCCESS, PAYMENT_FAILED, SHIPPED]
        subject:
          type: string
        bodyHtml:
          type: string
        availableVariables:
          type: array
          description: 可用變數清單（REQ-032）
          items:
            $ref: '#/components/schemas/TemplateVariableDTO'
        updatedAt:
          type: string
          format: date-time
    TemplateVariableDTO:
      type: object
      description: Email 模板可用變數定義（REQ-032）
      properties:
        variable:
          type: string
          description: 變數名稱（如 {{customerName}}）
          example: "{{customerName}}"
        description:
          type: string
          description: 變數中文說明
          example: "顧客姓名"
    UpdateEmailTemplateRequest:
      type: object
      required: [subject, bodyHtml]
      properties:
        subject:
          type: string
          minLength: 1
        bodyHtml:
          type: string
          minLength: 1
          description: 模板內文（HTML）。可用變數：{{customerName}}、{{orderNumber}}、{{totalAmount}}、{{websiteName}}、{{contactInfo}}（REQ-032）

    # ── Common ──
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time